+++ changelog.md
@@ -13,6 +13,7 @@
 ## Bug fixes
 
 * Fixed an issue where source and target IDs of relationships in heterogeneous OGBL graphs were not parsed correctly.
+* Fixed an issue where configuration parameters such as `aggregation` were not ignored by `gds.graph.toUndirected`.
 
 
 ## Improvements